# server/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.FilamentRequest:
    properties:
      bed_temp_max:
        example: 65
        type: integer
      bed_temp_min:
        example: 60
        type: integer
      color_hex:
        example: '#1a1a2e'
        type: string
      density:
        example: 1.24
        type: number
      dry_temp:
        example: 50
        type: integer
      dry_time:
        example: 6
        type: integer
      filled_type:
        example: Silk
        type: string
      id_brand:
        example: uuid-here
        type: string
      image_url:
        type: string
      material_type:
        example: PLA
        type: string
      nozzle_temp_max:
        example: 220
        type: integer
      nozzle_temp_min:
        example: 200
        type: integer
      title:
        example: Galaxy Black PLA
        type: string
      weight_grams:
        example: 1000
        type: integer
    type: object
...
  handlers.PreferenceRequest:
    properties:
      bed_temp_override:
        example: 60
        type: integer
      ironing_flow:
        example: 15
        type: integer
      ironing_speed:
        example: 30
        type: integer
      notes:
        example: prints better at 220
        type: string
      nozzle_temp_override:
        example: 215
        type: integer
    type: object
...
  models.Brand:
    properties:
      created_at:
        type: string
      empty_spool_weight_grams:
        type: number
      id_brand:
        type: string
      id_origin:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
  models.Filament:
    properties:
      bed_temp_max:
        type: integer
      bed_temp_min:
        type: integer
      brand:
        $ref: '#/definitions/models.Brand'
      color_hex:
        type: string
      color_info:
        type: object
      created_at:
        type: string
      density:
        type: number
      dry_temp:
        type: integer
      dry_time:
        type: integer
      filled_type:
        type: string
      id_brand:
        type: string
      id_filament:
        type: string
      id_origin:
        type: string
      image_url:
        type: string
      material_type:
        type: string
      nozzle_temp_max:
        type: integer
      nozzle_temp_min:
        type: integer
      product_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      weight_grams:
        type: integer
    type: object
  models.Preference:
    properties:
      bed_temp_override:
        type: integer
      id_filament:
        type: string
      id_user:
        type: string
      ironing_flow:
        type: integer
      ironing_speed:
        type: integer
      notes:
        type: string
      nozzle_temp_override:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Filament inventory management for 3D printing enthusiasts.
  title: Filaventory API
  version: "1.0"
paths:
...
  /api/filaments:
    get:
      parameters:
      - description: Full-text search (brand, title, material)
        in: query
        name: q
        type: string
      - collectionFormat: csv
        description: Filter by material types
        in: query
        items:
          type: string
        name: materials
        type: array
      - description: Filter by brand UUID
        in: query
        name: brand_id
        type: string
      - description: 'Sort field: title|brand|material|nozzle|bed|weight'
        in: query
        name: sort_by
        type: string
      - description: 'Sort direction: asc|desc'
        in: query
        name: sort_dir
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/models.Filament'
                type: array
              page:
                type: integer
              per_page:
                type: integer
              total:
                type: integer
            type: object
      security:
      - BearerAuth: []
      summary: List filaments
      tags:
      - filaments
    post:
      consumes:
      - application/json
      parameters:
      - description: Filament data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.FilamentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Filament'
        "400":
          description: validation error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a filament
      tags:
      - filaments
  /api/filaments/{id}:
    delete:
      parameters:
      - description: Filament UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
        "409":
          description: filament is in use by a spool
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a filament
      tags:
      - filaments
    get:
      parameters:
      - description: Filament UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Filament'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a filament
      tags:
      - filaments
    put:
      consumes:
      - application/json
      parameters:
      - description: Filament UUID
        in: path
        name: id
        required: true
        type: string
      - description: Filament data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.FilamentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Filament'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a filament
      tags:
      - filaments
  /api/filaments/materials:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List distinct material types
      tags:
      - filaments
...
  /api/preferences/{filament_id}:
    delete:
      parameters:
      - description: Filament UUID
        in: path
        name: filament_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete filament preference
      tags:
      - preferences
    get:
      parameters:
      - description: Filament UUID
        in: path
        name: filament_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Preference'
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get filament preference
      tags:
      - preferences
    put:
      consumes:
      - application/json
      parameters:
      - description: Filament UUID
        in: path
        name: filament_id
        required: true
        type: string
      - description: Preference data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.PreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Preference'
      security:
      - BearerAuth: []
      summary: Create or update filament preference
      tags:
      - preferences
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" in the Authorization header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"